+++ CMakeLists.txt
@@ -343,6 +343,7 @@ set(CMAKE_RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}")
 # FIXME: We should drop the dependency on ``mk_util.py``
 set(Z3_GENERATED_FILE_EXTRA_DEPENDENCIES
   "${CMAKE_SOURCE_DIR}/scripts/mk_util.py"
+  "${CMAKE_SOURCE_DIR}/scripts/mk_genfile_common.py"
 )
 
 ################################################################################
